flask_server/requirements.txt
Flask
outlines==1.0.4
python-dotenv
uvicorn
Pillow
numpy
ollama
pdf2image
transformers
torch
accelerate
datasets
torchvision
beautifulsoup4
pyjwt
celery
dash